pyproject.toml
[tool.poetry]
name = "tensorrt-yolo"
version = "0.1.0"
description = ""
authors = ["justin-prnd <justin@prnd.co.kr>"]
readme = "README.md"
packages = [{ include = "tensorrt_yolo" }]

[tool.poetry.dependencies]
python = "^3.8"
setuptools = "^67.2.0"
pip = "^23.0"
nvidia-tensorrt = { version = "^99.0.0", platform = 'linux' }
pycuda = { version = "^2022.2.2", platform = 'linux' }
matplotlib = "^3.6.3"
opencv-python = "^4.7.0.68"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
